+++ dispatch-service/src/main/java/de/muenchen/oss/swim/dispatcher/adapter/in/schedule/ScheduleAdapter.java
@@ -1,6 +1,7 @@
 package de.muenchen.oss.swim.dispatcher.adapter.in.schedule;
 
 import de.muenchen.oss.swim.dispatcher.application.port.in.DispatcherInPort;
+import de.muenchen.oss.swim.dispatcher.application.port.in.ProtocolProcessingInPort;
 import lombok.RequiredArgsConstructor;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
@@ -9,6 +10,7 @@ import org.springframework.stereotype.Component;
 @RequiredArgsConstructor
 public class ScheduleAdapter {
     private final DispatcherInPort dispatcherInPort;
+    private final ProtocolProcessingInPort protocolProcessingInPort;
 
     @Scheduled(cron = "${swim.dispatching-cron}")
     public void triggerDispatching() {
@@ -17,6 +19,6 @@ public class ScheduleAdapter {
 
     @Scheduled(cron = "${swim.protocol-processing-cron}")
     public void triggerProtocolProcessing() {
-        dispatcherInPort.triggerProtocolProcessing();
+        protocolProcessingInPort.triggerProtocolProcessing();
     }
 }
