+++ dispatch-service/src/main/java/de/muenchen/oss/swim/dispatcher/application/port/in/DispatcherInPort.java
@@ -6,10 +6,4 @@ public interface DispatcherInPort {
      * Loops over all configured use cases and triggers dispatching for matching files.
      */
     void triggerDispatching();
-
-    /**
-     * Trigger processing of protocol files.
-     * Loops over all configured use cases, loads, validates and processes matching protocol files.
-     */
-    void triggerProtocolProcessing();
 }
